eval/requirements.txt
numpy
scipy
matplotlib
pandas
pebble
tqdm
opencv-contrib-python==4.7.0.72

tensorflow==2.11.0
tensorflow-datasets==4.8.3
tensorflow-probability==0.19.0
tensorflow-addons==0.19.0
cupy-cuda11x

torch==2.0.1+cu118  # --index-url https://download.pytorch.org/whl/cu118
torchgeometry==0.1.2
torchmetrics==0.11.4
torchvision==0.15.2+cu118
segmentation-models-pytorch==0.3.2
einops==0.6.1

deeptrack==1.5.5